Implement getting and dismissing server announcements (#4)
to be fair, I didn't test it on a real mastodon server. I just hope it works. I also don't think it matters much anyway as this is a fork for calckey? idk please let me know if I need to test it on mastodon too.

It's a bit silly implementation as mfm is just escaped to plaintext and the announcement title is glued to `content`, but overall I think it works fine :)

Reviewed-on: https://codeberg.org/calckey/megalodon/pulls/4
Co-authored-by: fruye <fruye@unix.dog>
Co-committed-by: fruye <fruye@unix.dog>